requirements.txt
pysolarmanv5
python-dotenv
pytest
